+++ src/client/tcp.rs
@@ -1,4 +1,4 @@
-use super::Client;
+use super::Connection;
 
 use crate::service;
 
@@ -8,10 +8,10 @@ use std::net::SocketAddr;
 use tokio_core::reactor::Handle;
 
 pub fn connect(
-    socket_addr: &SocketAddr,
+    socket_addr: SocketAddr,
     handle: &Handle,
-) -> impl Future<Item = Client, Error = Error> {
-    service::tcp::Client::connect(socket_addr, handle).map(|service| Client {
+) -> impl Future<Item = Connection, Error = Error> {
+    service::tcp::Client::connect(&socket_addr, handle).map(|service| Connection {
         service: Box::new(service),
     })
 }