+++ src/codec/tcp.rs
@@ -8,6 +8,7 @@ use std::io::{Error, ErrorKind, Result};
 use tokio_codec::{Decoder, Encoder};
 
 const HEADER_LEN: usize = 7;
+
 const PROTOCOL_ID: u16 = 0x0000; // TCP
 
 #[derive(Debug, PartialEq)]